refactor: move window-manager runtime helpers into their own file
The runtime core crossed the 500-line production cap; the id generator,
the diagnostic notice TTL, and the query-cache change predicate now live
in window-manager-runtime-helpers.ts.
